+++ dart/lib/src/event_processor.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import '../protocol.dart';
+import 'protocol.dart';
 
 /// [EventProcessor]s are callbacks that run for every event. They can either
 /// return a new event which in most cases means just adding data *or* return