+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Enrich events with more context (#452)
+
 # 5.1.0
 
 * Fix: Merge user from event and scope (#467)