+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Remove `Sentry.currentHub` (#490)
+
 # 5.1.0
 
 * Fix: Merge user from event and scope (#467)