+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* Feat: EventProcessor changed to an interface (#489)
+* Ref: EventProcessor changed to an interface (#489)
 * Remove `Sentry.currentHub` (#490)
 
 # 5.1.0