+++ CHANGELOG.md
@@ -5,7 +5,14 @@
 * Feat: Support envelope based transport for events (#391)
   * The method signature of `Transport` changed from `Future<SentryId> send(SentryEvent event)` to `Future<SentryId> send(SentryEnvelope envelope)`
   * Bumped sentry-cocoa to v7.1.3
+* Remove `Sentry.currentHub` (#490)
+
+# 5.1.0
+
 * Fix: Merge user from event and scope (#467)
+* Feature: Allow setting of default values for in-app-frames via `SentryOptions.considerInAppFramesByDefault` (#482)
+* Bump: sentry-android to v5.0.1 (#486)
+* Bump: Sentry-Cocoa to 7.1.3 for iOS and macOS (#488)
 
 # 5.1.0-beta.1